refactor(agent): integrate agentClient for fetching agent input schema and job status, remove unused third-party functions
