+++ web-app/src/lib/clients/index.ts
@@ -1,1 +1,2 @@
+export * from "./agent.client";
 export * from "./stripe.client";
